Implement download users as CSV and delete all unrelevant js code.
